Efficient CapsNet/requirements.txt
numpy
pandas
tensorflow-addons
opencv-python
tqdm
matplotlib
pytest
tensorflow-datasets